Fix settings list background on iOS 15
