+++ apps/ios/GuideDogs/Code/Visual UI/Views/Settings/Language/LanguageSettingsView.swift
@@ -66,9 +66,8 @@ struct LanguageSettingsView: View {
                 }
             }
         }
-        .languageSettingsListBackground()
+        .settingsListBackground()
         .background(Color.quaternaryBackground.ignoresSafeArea())
-        .background(LanguageSettingsListBackgroundConfigurator())
         .listStyle(PlainListStyle())
         .tint(.primaryForeground)
         .navigationTitle(GDLocalizedString("settings.language.screen_title.2"))
@@ -234,74 +233,6 @@ private struct LanguageSettingsLocaleRow: View {
         .accessibilityElement(children: .combine)
         .accessibilityAddTraits(.isButton)
         .accessibilityAddTraits(isSelected ? .isSelected : [])
-    }
-}
-
-private extension View {
-    @ViewBuilder
-    func languageSettingsListBackground() -> some View {
-        if #available(iOS 16.0, *) {
-            scrollContentBackground(.hidden)
-        } else {
-            self
-        }
-    }
-}
-
-private struct LanguageSettingsListBackgroundConfigurator: UIViewRepresentable {
-    func makeUIView(context: Context) -> UIView {
-        let view = UIView(frame: .zero)
-        view.isHidden = true
-        clearTableBackground(from: view)
-        return view
-    }
-
-    func updateUIView(_ uiView: UIView, context: Context) {
-        clearTableBackground(from: uiView)
-    }
-
-    private func clearTableBackground(from view: UIView) {
-        guard #unavailable(iOS 16.0) else {
-            return
-        }
-
-        DispatchQueue.main.async {
-            view.enclosingTableView()?.backgroundColor = .clear
-        }
-    }
-}
-
-private extension UIView {
-    func enclosingTableView() -> UITableView? {
-        var currentView: UIView? = self
-
-        while let view = currentView {
-            if let tableView = view as? UITableView {
-                return tableView
-            }
-
-            if let tableView = view.firstSubview(of: UITableView.self) {
-                return tableView
-            }
-
-            currentView = view.superview
-        }
-
-        return nil
-    }
-
-    func firstSubview<T: UIView>(of type: T.Type) -> T? {
-        for subview in subviews {
-            if let matchingSubview = subview as? T {
-                return matchingSubview
-            }
-
-            if let matchingSubview = subview.firstSubview(of: type) {
-                return matchingSubview
-            }
-        }
-
-        return nil
     }
 }
